refactor 프로젝트 구조 ddd -> domain 계층 구조
