--- src/main/java/capstone/recipable/domain/user/domain/Provider.java
@@ -1,5 +1,0 @@
-package capstone.recipable.domain.user.domain;
-
-public enum Provider {
-    KAKAO
-}
