--- src/main/java/capstone/recipable/domain/user/domain/oauth/OAuthHandler.java
@@ -1,5 +1,0 @@
-package capstone.recipable.domain.user.domain.oauth;
-
-public interface OAuthHandler {
-    OAuthTransactionResult retrieveOAuthDetail(OAuthProcessingData request);
-}
