--- src/main/java/capstone/recipable/domain/user/domain/oauth/OAuthProcessingData.java
@@ -1,6 +1,0 @@
-package capstone.recipable.domain.user.domain.oauth;
-
-public record OAuthProcessingData(
-        String accessToken
-) {
-}
