--- src/main/java/capstone/recipable/domain/user/domain/oauth/OAuthTransactionResult.java
@@ -1,6 +1,0 @@
-package capstone.recipable.domain.user.domain.oauth;
-
-public record OAuthTransactionResult(
-        String sub
-) {
-}
